+++ CHANGELOG.md
@@ -57,8 +57,10 @@ to geth v1.14 with tracing updates and new StateDB methods.
 - [#2292](https://github.com/NibiruChain/nibiru/pull/2292) - fix: use tmp directory for pre-instantiating app
 - [#2293](https://github.com/NibiruChain/nibiru/pull/2293) - ci(release): pack nibid binary with no enclosing directory
 - [#2296](https://github.com/NibiruChain/nibiru/pull/2296) - chore(ci): use shell script for generating changelog in releases
+- [#2297](https://github.com/NibiruChain/nibiru/pull/2297) - fix(evm): fix error handling for revert errors
 
 ### Dependencies
+
 - Bump `golang.org/x/net` from 0.37.0 to 0.39.0. ([#2284](https://github.com/NibiruChain/nibiru/pull/2284))
 - Bump `github.com/golang-jwt/jwt/v4` from 4.5.1 to 4.5.2 ([#2294](https://github.com/NibiruChain/nibiru/pull/2294))